refactor(data): rework vehicle location service based on new knowledge
The library I was using did not work they way I thought it was going to. After doing some digging, it turns out to not be a very good wrapper for my needs and I am better off using the Google Maps API directly. That requires a slight change in data structures.